Payments/requirements.txt
boto3
requests
stripe
python-dotenv

